+++ src/components/Form/Select/constants.ts
@@ -1,8 +1,6 @@
 import { CommonAndHTMLProps } from "@/components/Element/constants";
-import React from "react";
 
 export type SelectElementType = HTMLSelectElement;
-export type SelectWithSearchElementType = HTMLSelectElement;
 
 export type OptionElementType = HTMLOptionElement;
 
@@ -11,13 +9,6 @@ export interface OptionProps {
     label      : string;
     disabled ? : boolean;
     selected ? : boolean;
-}
-
-export interface OptionForSearchWithSelectProps {
-    value         : string;
-    label         : string;
-    customLabel ? : React.ReactNode;
-    disabled    ? : boolean;
 }
 
 export interface OptGroupProps {
@@ -32,17 +23,4 @@ export interface SelectCustomProps {
     errorText ? : string;
 }
 
-export interface SelectWithSearchCustomProps {
-    options     : OptionForSearchWithSelectProps[];
-    label     ? : string;
-    helpText  ? : string;
-    errorText ? : string;
-}
-
-export interface SelectWithSearchCustomProps extends SelectCustomProps {
-    onChange     ? : (value: string) => void;
-    defaultValue ? : string;
-}
-
 export type SelectProps = Omit<CommonAndHTMLProps<SelectElementType>, keyof SelectCustomProps> & SelectCustomProps;
-export type SelectWithSearchProps = CommonAndHTMLProps<SelectWithSearchElementType> & SelectWithSearchCustomProps;