+++ src/components/Form/index.tsx
@@ -1,5 +1,8 @@
 export { Checkbox, type CheckboxProps } from "./Checkbox/Checkbox";
 
+export { ListBox } from "./ListBox/ListBox";
+export { type ListBoxProps } from "./ListBox/constants";
+
 export { FileUpload, type FileUploadProps } from "./InputField/FileUpload";
 
 export { Form, type FormProps } from "./Form/Form";
@@ -22,8 +25,7 @@ export { type RadioButtonProps, type RadioGroupProps, type RadioTabGroupProps } 
 export { Range, type RangeProps } from "./Range";
 
 export { Select } from "./Select/Select";
-export { SelectWithSearch } from "./Select/SelectWithSearch";
-export { type SelectProps, type SelectWithSearchProps } from "./Select/constants";
+export { type SelectProps } from "./Select/constants";
 
 export { RadioTabGroup } from "./RadioButton/RadioTabGroup";