+++ src/components/index.tsx
@@ -60,7 +60,6 @@ export {
     type InputLabelProps,
     Select,
     type SelectProps,
-    SelectWithSearch,
     FileUpload,
     type FileUploadProps,
     RadioGroup,
@@ -77,6 +76,8 @@ export {
     type RadioTabGroupProps,
     PinInputField,
     type PinInputFieldProps,
+    ListBox,
+    type ListBoxProps
 } from "./Form";
 
 // METER ===============================================================================================================